requirements-dev.txt
astroid==2.1.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
six==1.11.0
static3==0.7.0
wrapt==1.10.11
